Improve the speed of converting flac files
Improve the speed of converting flac files for large files.
"When used as an input option (before -i), seeks in this input file to position. When used as an output option (before an output filename), decodes but discards input until the timestamps reach position. This is slower, but more accurate."

http://stackoverflow.com/questions/6984628/ffmpeg-ss-weird-behaviour.
For Movies 2 hrs, is notable performance.

3-10 minutes vs 2 seg For convert flac.